돌보미 등록 적용, 회원가입 시 돌보미 등록 개발, 기존 돌보미가 아닌경우 돌보미 등록, 돌보미 인 경우 수정 기능 적용, 돌보미 목록 조회 화면 개발 중

+++ src/main/java/sch/pl/graduate/recommendation/user/common/controller/UserController.java
@@ -21,7 +21,7 @@ public class UserController extends AbstractController {
 
     @PostMapping("/users")
     public ResponseEntity addUser(@RequestBody User user) {
-        Integer result = userService.addUser(user);
+        Long result = userService.addUser(user);
 
         return getSuccessResponse(result, "성공하였습니다");
     }
